// src/app/seller/interfaces/Information.interface.ts
export interface InformationInterface {
    uid: string;
    country:     string;
    costumerId:  string;
    firstName:   string;
    lastName:    string;
    birthday:    Date;
    gender:      string;
    cellPhone:   string;
    homePhone:   string;
    addressHome: string;
    profession:  string;
    incomes:     number;
}